# Docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Application Modernization Web App REST API",
    "description": "Backend REST API for the Application Modernization",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhot:8080/api/ver{version}/",
      "description": "Local Development",
      "variables": {
        "version": {
          "default": "1",
          "description": "The API version number"
        }
      }
    }
  ],
  "paths": {
...
    "/views/{id}": {
      "delete": {
        "tags": [
          "View"
        ],
        "summary": "Deletes View with the given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The View Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The View Id.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "View"
        ],
        "summary": "Get View with the given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "View"
        ],
        "summary": "Updates View with the given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The view Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The view Id.",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The view model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
    "/views/{viewId}/rules/{ruleId}": {
      "delete": {
        "tags": [
          "View"
        ],
        "summary": "Deletes View Rule with given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "viewId",
            "in": "path",
            "description": "The view Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The view Id.",
              "format": "int32"
            }
          },
          {
            "name": "ruleId",
            "in": "path",
            "description": "The rule Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The rule Id.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "View"
        ],
        "summary": "Gets View Rule with given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "viewId",
            "in": "path",
            "description": "The view Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The view Id.",
              "format": "int32"
            }
          },
          {
            "name": "ruleId",
            "in": "path",
            "description": "The rule Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The rule Id.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewRuleModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "View"
        ],
        "summary": "Updates View Rule with given Id.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "viewId",
            "in": "path",
            "description": "The view Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The view Id.",
              "format": "int32"
            }
          },
          {
            "name": "ruleId",
            "in": "path",
            "description": "The rule Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The rule Id.",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The View Rule model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewRuleModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewRuleModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewRuleModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
...
    "/views/fields/{fieldId}/values": {
      "get": {
        "tags": [
          "View"
        ],
        "summary": "Get list of values for the field with SelectionType='values'.",
        "description": "<p></p>",
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldValueModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
    "/views": {
      "get": {
        "tags": [
          "View"
        ],
        "summary": "Get all current user's Views.",
        "description": "<p></p>",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ViewModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "View"
        ],
        "summary": "Creates new View.",
        "description": "<p></p>",
        "requestBody": {
          "description": "The view model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Gets or sets the message.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "An API error Model."
      },
...
      "ViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The Id of the model.",
            "format": "int32",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "The view name.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "The view description.",
            "nullable": true
          },
          "limits": {
            "$ref": "#/components/schemas/ViewLimitsModel"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Indicating whether this view is default."
          }
        },
        "additionalProperties": false,
        "description": "The View model."
      },
      "CreateViewModel": {
        "required": [
          "description",
          "isDefault",
          "limits",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The view name."
          },
          "description": {
            "type": "string",
            "description": "The view description."
          },
          "limits": {
            "$ref": "#/components/schemas/ViewLimitsModel"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Indicating whether this view is default."
          }
        },
        "additionalProperties": false,
        "description": "The Create View model."
      },
      "ViewRuleModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The Id of the model.",
            "format": "int32",
            "example": 1
          },
          "fieldId": {
            "type": "integer",
            "description": "The field Id.",
            "format": "int32"
          },
          "fieldName": {
            "type": "string",
            "description": "The field Name.",
            "nullable": true,
            "example": "Payer1AuditFlag"
          },
          "value": {
            "type": "string",
            "description": "The Value.",
            "nullable": true,
            "example": "My First Contract"
          },
          "valueId": {
            "type": "string",
            "description": "The selected Id in case field type is multi-select.",
            "nullable": true
          },
          "beginRange": {
            "type": "string",
            "description": "The begin range.",
            "nullable": true,
            "example": "01/01/1900"
          },
          "endRange": {
            "type": "string",
            "description": "The end range.",
            "nullable": true,
            "example": "12/31/1986"
          },
          "operand": {
            "type": "string",
            "description": "The operand.",
            "nullable": true,
            "example": "BETWEEN"
          }
        },
        "additionalProperties": false,
        "description": "The View model."
      },
...
    },
    "securitySchemes": {
      "jwt_bearer": {
        "type": "http",
        "description": "Authentication using JWT Token in header.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}